Refactor: Signal handler moved to views
