+++ backend/api/views.py
@@ -27,7 +27,10 @@ from django.core.mail import send_mail
 from templates.consts import *
 
 from rest_framework import status
-from rest_framework.permissions import IsAuthenticated  
+from rest_framework.permissions import IsAuthenticated
+
+from django.dispatch import receiver
+from django_rest_passwordreset.signals import reset_password_token_created
 
 
 def get_wss_object_or_404(year: int) -> WSS:
@@ -482,3 +485,20 @@ class ChangePasswordView(generics.UpdateAPIView):
             }, status=status.HTTP_200_OK)
 
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+''' Signal handlers should be placed somewhere which is automatically loaded (like here or models)
+or be imported in ready() function in apps.py '''
+@receiver(reset_password_token_created)
+def password_reset_token_created(sender, instance, reset_password_token, *args, **kwargs):
+
+    base_url = 'https://sharif-wss.ir/'  # TODO
+    email_message = "{}?token={}".format(base_url, reset_password_token.key)
+
+    send_mail(
+        RESET_PASSWORD_SUBJECT, 'text content',
+        settings.EMAIL_HOST_USER,
+        [reset_password_token.user.email],
+        fail_silently=True,
+        html_message=BASE_HTML_CONTENT.format(RESET_PASSWORD_EMAIL.format(email_message))
+    )
